Clientlayoutコンテンツのコンポーネント化
