+++ app/layout.tsx
@@ -1,4 +1,5 @@
 import { ReactNode } from "react";
+import {Staatliches} from "next/font/google"
 import { Providers } from "./providers";
 import ClientLayout from "./Clientlayout";
 
@@ -7,18 +8,21 @@ export const metadata = {
   description: "MIYA Guide - 自作PCガイド",
 };
 
+const staatliches = Staatliches({
+  weight: '400',
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-staatliches',
+})
+
 type LayoutProps = {
   children: ReactNode;
 };
 
 export default function RootLayout({ children }: LayoutProps) {
   return (
-    <html lang="ja">
+    <html lang="ja" className={staatliches.variable}>
       <head>
-        <link
-          href="https://fonts.googleapis.com/css2?family=Staatliches&display=swap"
-          rel="stylesheet"
-        />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </head>
       <body>
