+++ components/organisms/PartsCard.tsx
@@ -74,7 +74,7 @@ const PartsCard: React.FC<PartsCardProps> = ({
   return (
     <MotionCard
       variant="neon"
-      initial={{ opacity: 0, y: 20 }}
+      initial={false}
       animate={{ opacity: 1, y: 0 }}
       transition={{
         duration: 0.5,
